requirements.txt
asgiref==3.4.1
certifi==2021.5.30
charset-normalizer==2.0.3
dj-rest-auth==2.1.9
Django==3.2.5
djangorestframework==3.12.4
idna==3.2
psycopg2==2.9.1
pycodestyle==2.7.0
pytz==2021.1
requests==2.26.0
sqlparse==0.4.1
urllib3==1.26.6
